Separate decimals (dal, node) (#2486)
* feat: support decimals

* chore: minor fix

* fix: modify dal

* fix: fix nil map err

* fix: update logs based on feedback

* chore: remove not null condition

* fix: fix test

* test: update test based on updated logic

+++ node/pkg/fetcher/localaggregatebulkwriter.go
@@ -44,7 +44,7 @@ loop:
 	for {
 		select {
 		case data := <-a.localAggregatesChannel:
-			localAggregatesDataPgsql = append(localAggregatesDataPgsql, []any{data.ConfigID, int64(data.Value), data.Timestamp})
+			localAggregatesDataPgsql = append(localAggregatesDataPgsql, []any{data.ConfigID, data.Value, data.Timestamp})
 		default:
 			break loop
 		}
